+++ djpcms/middleware.py
@@ -12,4 +12,5 @@ class DjpcmsRequestHandler(object):
     def process_request(self, request):
         pass
     
+